[Fixes #1305] Unable to upload GEOJSON files with .json extention (#1317)
@marthamareal please test on master and let us know if we can backport, thanks
