+++ geonode_mapstore_client/client/js/routes/UploadDataset.jsx
@@ -25,17 +25,10 @@ import UploadListContainer from '@js/routes/upload/UploadListContainer';
 import UploadContainer from '@js/routes/upload/UploadContainer';
 import { getConfigProp } from '@mapstore/framework/utils/ConfigUtils';
 import { parseUploadResponse, processUploadResponse, parseUploadFiles } from '@js/utils/ResourceUtils';
-
-function getFileNameParts(file) {
-    const { name } = file;
-    const nameParts = name.split('.');
-    const ext = nameParts[nameParts.length - 1];
-    const baseName = [...nameParts].splice(0, nameParts.length - 1).join('.');
-    return { ext, baseName };
-}
+import { getFileNameParts, getFileType } from '@js/utils/FileUtils';
 
 function getDatasetFileType(file, supportedTypes) {
-    const { type } = file;
+    const type = getFileType(file);
     const { ext } = getFileNameParts(file);
     const datasetFileType = supportedTypes.find((fileType) =>
         (fileType.ext || []).includes(ext)
